Feature/Code Formatting (#19) * rename root package to eft
