+++ .run/ExamFeedbackToolApplication.run.xml
@@ -4,7 +4,7 @@
       <env name="RESOURCE_DIR" value="./static" />
     </envs>
     <module name="exam-feedback-tool" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="ch.puzzle.exam_feedback_tool.ExamFeedbackToolApplication" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="ch.puzzle.eft.ExamFeedbackToolApplication" />
     <method v="2">
       <option name="Make" enabled="true" />
     </method>
