+++ src/main/java/ch/puzzle/eft/ExamFeedbackToolApplication.java
@@ -1,4 +1,4 @@
-package ch.puzzle.exam_feedback_tool;
+package ch.puzzle.eft;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
