+++ src/test/java/ch/puzzle/eft/controller/DemoControllerTest.java
@@ -1,7 +1,7 @@
-package ch.puzzle.exam_feedback_tool.controller;
+package ch.puzzle.eft.controller;
 
 
-import ch.puzzle.exam_feedback_tool.service.DemoService;
+import ch.puzzle.eft.service.DemoService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -9,7 +9,6 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
