+++ src/test/java/ch/puzzle/eft/service/DemoServiceTest.java
@@ -1,4 +1,4 @@
-package ch.puzzle.exam_feedback_tool.service;
+package ch.puzzle.eft.service;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
